network dns hacks: split from network.py into its own file
